fix(catalog): address review feedback
## Description

- Wire the Custom rate model write path: extract the custom-charge sub-drawer from
  UsageChargeDrawerContent into a shared useCustomChargeDrawer hook and pass
  onExpandCustomCharge from the rate drawer, so the typed JSON reaches the form
- Replace the RateCardRateModelEnum -> ChargeModelEnum casts with an exhaustive
  toChargeModel record, so a future divergence between the two enums fails to compile
- Narrow ChargeModelSelector / ChargeWrapperSwitch / SpendingMinimumOptionSection
  localCharge props to the single field each reads, removing the synthetic
  LocalUsageChargeInput casts in the rate drawer and the rateProperties cast on the
  rate overview
- Split drawers/rateCardRate: constants.ts now holds only constants, with schema.ts,
  utils.ts and mapRateToFormValues.ts alongside it, and the form hook moved out of
  useRateCardRateDrawer.tsx into useRateCardRateForm.tsx
- Move escapeDoubleQuotes back to ~/core/utils (nothing about it is catalog-specific)
- Drop the repeated translation-key banner comment and trim the prop / cache comments
  flagged as noise; drop the BIL-594 reference from the activity logs tab
- Unit tests for the new utils, schema and mapRateToFormValues modules, plus a
  regression test pinning the Custom rate model write path

<!-- Linear link -->
Refs BIL-501
