+++ src/core/utils/__tests__/escapeDoubleQuotes.test.ts
