+++ src/core/utils/escapeDoubleQuotes.ts
